Refactored tie breaking functionality into dedicated namespace.
